+++ test/foundry/modules/licensing/PILicenseTemplate.t.sol
@@ -52,7 +52,7 @@ contract PILicenseTemplateTest is BaseTest {
     // this contract is for testing for each PILicenseTemplate's functions
     // register license terms with PILTerms struct
     function test_PILicenseTemplate_registerLicenseTerms() public {
-        uint256 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         assertEq(defaultTermsId, 1);
         (address royaltyPolicy, bytes memory royaltyData, uint256 mintingFee, address currency) = pilTemplate
             .getRoyaltyPolicy(defaultTermsId);
@@ -65,7 +65,7 @@ contract PILicenseTemplateTest is BaseTest {
         assertEq(pilTemplate.getExpireTime(defaultTermsId, block.timestamp), 0, "expire time should be 0");
         assertTrue(pilTemplate.exists(defaultTermsId), "license terms should exist");
 
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         assertEq(socialRemixTermsId, 2);
         (royaltyPolicy, royaltyData, mintingFee, currency) = pilTemplate.getRoyaltyPolicy(socialRemixTermsId);
         assertEq(royaltyPolicy, address(0));
@@ -77,7 +77,7 @@ contract PILicenseTemplateTest is BaseTest {
         assertEq(pilTemplate.getExpireTime(socialRemixTermsId, block.timestamp), 0, "expire time should be 0");
         assertTrue(pilTemplate.exists(socialRemixTermsId), "license terms should exist");
 
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -98,7 +98,7 @@ contract PILicenseTemplateTest is BaseTest {
         assertEq(pilTemplate.getExpireTime(commUseTermsId, block.timestamp), 0, "expire time should be 0");
         assertEq(pilTemplate.totalRegisteredLicenseTerms(), 3);
 
-        uint256 commRemixTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commRemixTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialRemix({
                 mintingFee: 100,
                 commercialRevShare: 10,
@@ -124,7 +124,7 @@ contract PILicenseTemplateTest is BaseTest {
 
         assertEq(pilTemplate.totalRegisteredLicenseTerms(), 4);
 
-        uint256[] memory licenseTermsIds = new uint256[](4);
+        uint32[] memory licenseTermsIds = new uint32[](4);
         licenseTermsIds[0] = defaultTermsId;
         licenseTermsIds[1] = socialRemixTermsId;
         licenseTermsIds[2] = commUseTermsId;
@@ -135,8 +135,8 @@ contract PILicenseTemplateTest is BaseTest {
     }
     // register license terms twice
     function test_PILicenseTemplate_registerLicenseTerms_twice() public {
-        uint256 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
-        uint256 defaultTermsId1 = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 defaultTermsId1 = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         assertEq(defaultTermsId, defaultTermsId1);
     }
 
@@ -233,7 +233,7 @@ contract PILicenseTemplateTest is BaseTest {
 
     // get license terms ID by PILTerms struct
     function test_PILicenseTemplate_getLicenseTermsId() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -253,7 +253,7 @@ contract PILicenseTemplateTest is BaseTest {
 
     // get license terms struct by ID
     function test_PILicenseTemplate_getLicenseTerms() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -268,7 +268,7 @@ contract PILicenseTemplateTest is BaseTest {
 
     // test license terms exists
     function test_PILicenseTemplate_exists() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -281,7 +281,7 @@ contract PILicenseTemplateTest is BaseTest {
 
     // test verifyMintLicenseToken
     function test_PILicenseTemplate_verifyMintLicenseToken() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -293,7 +293,7 @@ contract PILicenseTemplateTest is BaseTest {
     }
 
     function test_PILicenseTemplate_verifyMintLicenseToken_FromDerivativeIp_ButNotAttachedLicense() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 0,
                 currencyToken: address(erc20),
@@ -305,19 +305,19 @@ contract PILicenseTemplateTest is BaseTest {
 
         address[] memory parentIpIds = new address[](1);
         parentIpIds[0] = ipAcct[1];
-        uint256[] memory licenseTermsIds = new uint256[](1);
+        uint32[] memory licenseTermsIds = new uint32[](1);
         licenseTermsIds[0] = commUseTermsId;
         vm.prank(ipOwner[2]);
         licensingModule.registerDerivative(ipAcct[2], parentIpIds, licenseTermsIds, address(pilTemplate), "");
 
-        uint256 anotherTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 anotherTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
 
         bool result = pilTemplate.verifyMintLicenseToken(anotherTermsId, ipOwner[3], ipAcct[2], 1);
         assertFalse(result);
     }
 
     function test_PILicenseTemplate_verifyMintLicenseToken_FromDerivativeIp_NotReciprocal() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 0,
                 currencyToken: address(erc20),
@@ -329,7 +329,7 @@ contract PILicenseTemplateTest is BaseTest {
 
         address[] memory parentIpIds = new address[](1);
         parentIpIds[0] = ipAcct[1];
-        uint256[] memory licenseTermsIds = new uint256[](1);
+        uint32[] memory licenseTermsIds = new uint32[](1);
         licenseTermsIds[0] = commUseTermsId;
         vm.prank(ipOwner[2]);
         licensingModule.registerDerivative(ipAcct[2], parentIpIds, licenseTermsIds, address(pilTemplate), "");
@@ -340,7 +340,7 @@ contract PILicenseTemplateTest is BaseTest {
 
     // test verifyRegisterDerivative
     function test_PILicenseTemplate_verifyRegisterDerivative() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -355,7 +355,7 @@ contract PILicenseTemplateTest is BaseTest {
     function test_PILicenseTemplate_verifyRegisterDerivative_WithApproval() public {
         PILTerms memory terms = PILFlavors.nonCommercialSocialRemixing();
         terms.derivativesApproval = true;
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipAcct[1]);
         pilTemplate.setApproval(ipAcct[1], socialRemixTermsId, ipAcct[2], true);
         assertTrue(pilTemplate.isDerivativeApproved(ipAcct[1], socialRemixTermsId, ipAcct[2]));
@@ -367,7 +367,7 @@ contract PILicenseTemplateTest is BaseTest {
     function test_PILicenseTemplate_verifyRegisterDerivative_WithoutApproval() public {
         PILTerms memory terms = PILFlavors.nonCommercialSocialRemixing();
         terms.derivativesApproval = true;
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipAcct[1]);
         pilTemplate.setApproval(ipAcct[1], socialRemixTermsId, ipAcct[2], false);
         assertFalse(pilTemplate.isDerivativeApproved(ipAcct[1], socialRemixTermsId, ipAcct[2]));
@@ -378,21 +378,21 @@ contract PILicenseTemplateTest is BaseTest {
 
     function test_PILicenseTemplate_verifyRegisterDerivative_derivativeNotAllowed() public {
         PILTerms memory terms = PILFlavors.defaultValuesLicenseTerms();
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(terms);
         bool result = pilTemplate.verifyRegisterDerivative(ipAcct[2], ipAcct[1], socialRemixTermsId, ipOwner[2]);
         assertFalse(result);
     }
 
     // test verifyCompatibleLicenses
     function test_PILicenseTemplate_verifyCompatibleLicenses() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
                 royaltyPolicy: address(royaltyPolicyLAP)
             })
         );
-        uint256 commRemixTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commRemixTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialRemix({
                 mintingFee: 100,
                 commercialRevShare: 10,
@@ -400,13 +400,13 @@ contract PILicenseTemplateTest is BaseTest {
                 currencyToken: address(erc20)
             })
         );
-        uint256[] memory licenseTermsIds = new uint256[](2);
+        uint32[] memory licenseTermsIds = new uint32[](2);
         licenseTermsIds[0] = commUseTermsId;
         licenseTermsIds[1] = commRemixTermsId;
         assertFalse(pilTemplate.verifyCompatibleLicenses(licenseTermsIds));
 
-        uint256 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         licenseTermsIds[0] = defaultTermsId;
         licenseTermsIds[1] = socialRemixTermsId;
         assertFalse(pilTemplate.verifyCompatibleLicenses(licenseTermsIds));
@@ -422,14 +422,14 @@ contract PILicenseTemplateTest is BaseTest {
 
     // test verifyRegisterDerivativeForAllParents
     function test_PILicenseTemplate_verifyRegisterDerivativeForAllParents() public {
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
                 royaltyPolicy: address(royaltyPolicyLAP)
             })
         );
-        uint256 commRemixTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commRemixTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialRemix({
                 mintingFee: 100,
                 commercialRevShare: 10,
@@ -437,7 +437,7 @@ contract PILicenseTemplateTest is BaseTest {
                 currencyToken: address(erc20)
             })
         );
-        uint256[] memory licenseTermsIds = new uint256[](2);
+        uint32[] memory licenseTermsIds = new uint32[](2);
         licenseTermsIds[0] = commUseTermsId;
         licenseTermsIds[1] = commRemixTermsId;
         address[] memory parentIpIds = new address[](2);
@@ -447,8 +447,8 @@ contract PILicenseTemplateTest is BaseTest {
             pilTemplate.verifyRegisterDerivativeForAllParents(ipAcct[2], parentIpIds, licenseTermsIds, ipOwner[2])
         );
 
-        uint256 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         licenseTermsIds[0] = defaultTermsId;
         licenseTermsIds[1] = socialRemixTermsId;
         assertFalse(
@@ -470,13 +470,13 @@ contract PILicenseTemplateTest is BaseTest {
 
     // test isLicenseTransferable
     function test_PILicenseTemplate_isLicenseTransferable() public {
-        uint256 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 defaultTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         assertTrue(pilTemplate.isLicenseTransferable(defaultTermsId));
 
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         assertTrue(pilTemplate.isLicenseTransferable(socialRemixTermsId));
 
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 100,
                 currencyToken: address(erc20),
@@ -485,7 +485,7 @@ contract PILicenseTemplateTest is BaseTest {
         );
         assertTrue(pilTemplate.isLicenseTransferable(commUseTermsId));
 
-        uint256 commRemixTermsId = pilTemplate.registerLicenseTerms(
+        uint32 commRemixTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialRemix({
                 mintingFee: 100,
                 commercialRevShare: 10,
@@ -497,7 +497,7 @@ contract PILicenseTemplateTest is BaseTest {
 
         PILTerms memory terms = pilTemplate.getLicenseTerms(commRemixTermsId);
         terms.transferable = false;
-        uint256 nonTransferableTermsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 nonTransferableTermsId = pilTemplate.registerLicenseTerms(terms);
         assertFalse(pilTemplate.isLicenseTransferable(nonTransferableTermsId));
     }
 
@@ -508,7 +508,7 @@ contract PILicenseTemplateTest is BaseTest {
             royaltyPolicy: address(royaltyPolicyLAP)
         });
         terms.uri = "license.url";
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         assertEq(pilTemplate.getLicenseTermsURI(termsId), "license.url");
     }
 
@@ -519,7 +519,7 @@ contract PILicenseTemplateTest is BaseTest {
             royaltyPolicy: address(royaltyPolicyLAP)
         });
         terms.uri = "license.url";
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         assertEq(pilTemplate.getLicenseTermsURI(termsId), "license.url");
 
         PILTerms memory terms1 = PILFlavors.commercialUse({
@@ -528,7 +528,7 @@ contract PILicenseTemplateTest is BaseTest {
             royaltyPolicy: address(royaltyPolicyLAP)
         });
         terms1.uri = "another.license.url";
-        uint256 termsId1 = pilTemplate.registerLicenseTerms(terms1);
+        uint32 termsId1 = pilTemplate.registerLicenseTerms(terms1);
 
         assertEq(pilTemplate.getLicenseTermsURI(termsId1), "another.license.url");
         assertEq(pilTemplate.getLicenseTermsURI(termsId), "license.url");
@@ -537,7 +537,7 @@ contract PILicenseTemplateTest is BaseTest {
     }
 
     function test_PILicenseTemplate_getEarlierExpiredTime_WithEmptyLicenseTerms() public {
-        uint256[] memory licenseTermsIds = new uint256[](0);
+        uint32[] memory licenseTermsIds = new uint32[](0);
         assertEq(pilTemplate.getEarlierExpireTime(licenseTermsIds, block.timestamp), 0);
     }
